TYPE FIXES

+++ src/main/java/com/example/test/repo/ValueRepo.java
@@ -3,6 +3,6 @@ package com.example.test.repo;
 import com.example.test.entity.Value;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ValueRepo extends JpaRepository<Value,Long> {
+public interface ValueRepo extends JpaRepository<Value,Integer> {
 
 }
